requirements.txt
fastapi
uvicorn
python-multipart
cloudinary
python-dotenv
bson
pymongo
motor
boto3
Pillow
httpx